ROX-17689: Move policy error definitions to datastore from bolt store
